feature/user-settings (#5)
* Add User CRUD + SettingController

* Make getUsername return email

* Rename variables

* Update file name, env variable, format code

* Update SecurityConfig add POST_WHITELIST

* Remove users username from migration

* Add KOL settings

* Update email url

* Update SecurityConfig add auth route

* Update images path to SecurityConfig

* Add resource + profile to pom.xml

* Add ResourceHandler -> WebConfig

* Add image resource -> WebConfig

* Add UserSetting CRUD, update flyway migration

+++ src/main/java/com/dtu/kolgo/model/Feedback.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class Feedback extends BaseInt {
 
     @Column
-    private int rating;
+    private short rating;
     @Column
     private String comment;
     @ManyToOne
